backend/pyproject.toml
[project]
name = "health-assistant-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
    "langgraph>=1.0.10",
    "chromadb>=0.4.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "openai>=1.0.0",
    "neo4j>=5.0.0",
    "pyyaml>=6.0",
    "redis>=5.0.0",
    "alibabacloud-oss-v2>=1.2.0",
    "alibabacloud-sts20150401>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
select = ["E", "F", "I"]
ignore = []
# E402：模块级 import 必须在文件顶部

[tool.ruff.lint.isort]
known-first-party = ["app"]
